modularizacion de las llamadas a la api
