+++ src/App.css
@@ -79,6 +79,7 @@ body {
   justify-content: center;
   align-items: center;
   font-weight: 600;
+  font-size: 18px;
 }
 .pokemon__descripcion {
   font-size: 18px;
