+++ src/api/obtenerEvoluciones.js
@@ -1,0 +1,17 @@
+const obtenerEvoluciones = (cadena) => {
+  let evoluciones = [];
+  let cadenaActual = cadena;
+
+  while (cadenaActual) {
+    evoluciones.push({
+      id: cadenaActual.species.url.split('/').slice(-2, -1)[0],
+      nombre: cadenaActual.species.name
+    });
+    cadenaActual = cadenaActual.evolves_to[0];
+  }
+
+  return evoluciones;
+};
+
+
+  export default obtenerEvoluciones;
